最新Verの処理と統合
