+++ .prompt/20251122.txt
@@ -7,3 +7,4 @@ GitHub Copilot [Agent] - Claude Sonnet 4.5
 - タイムラインビューアーで音楽ファイルの長さが、Telloの動きの秒数までしかないのを改善して。
 - タイムラインのズームの倍率を減らすと秒数の字が重なってしまうので、5秒おきや10秒おきに表示がかわるように改善
 - タイムラインウィンドウが複数開けないようにして
+- フォルダ"2025_10_27"の中にあるpythonコードが本システムの最新バージョンです。ただ一部の機能は先ほどこちらで上位互換を実装しました。特に音源の再生（メドレー機能）やタイムラインビューアーです。その他のTello_controllerたScratch_parser、show_runnerは本システムが古いバージョンなので"2025_10_27"内のデータを統合してください。
