+++ old-ver/2025_10_27/tello_config.json
@@ -1,0 +1,4 @@
+{
+    "Tello_A": "192.168.10.3",
+    "Tello_B": "192.168.10.2"
+}
